[open-formulieren/open-forms#5932] add confirmation email checkbox for profile component
