+++ package.json
@@ -100,6 +100,7 @@
     "prop-types": "^15.8.1",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
+    "react-select-event": "^5.5.1",
     "sass": "^1.75.0",
     "sass-loader": "^14.2.0",
     "storybook": "9.1.19",
@@ -117,7 +118,7 @@
     "@floating-ui/react": "^0.26.4",
     "@open-formulieren/leaflet-tools": "^1.0.0",
     "@open-formulieren/monaco-json-editor": "^0.2.0",
-    "@open-formulieren/types": "^1.0.0",
+    "@open-formulieren/types": "^1.1.0",
     "ckeditor5": "^47.6.1",
     "clsx": "^1.2.1",
     "formik": "^2.4.5",
